Use groovy library instead of module
Adds defaultSujbect as a possibility when subject key is not found
