+++ grails-app/init/grails/plugin/awssdk/ses/Application.groovy
