--- groovy-module/build.gradle
@@ -1,19 +1,0 @@
-apply plugin: 'groovy'
-
-repositories {
-    jcenter()
-}
-
-dependencies {
-    compile "org.codehaus.groovy:groovy-all:$groovyVersion"
-    compile "org.slf4j:slf4j-api:$slf4jVersion"
-    compile "ch.qos.logback:logback-classic:$logbackVersion"
-    compile "ch.qos.logback:logback-core:$logbackVersion"
-
-    compile "com.amazonaws:aws-java-sdk-ses:$awsJavaSdkVersion"
-
-    // Necessary to support attachments
-    compile 'javax.mail:mail:1.4.4'
-
-    testCompile "org.spockframework:spock-core:1.0-groovy-2.4"
-}
