--- groovy-module/src/main/groovy/grails/plugin/awssdk/ses/AwsSdkSesEmailDeliveryStatus.groovy
@@ -1,10 +1,0 @@
-package grails.plugin.awssdk.ses
-
-import groovy.transform.CompileStatic
-
-@CompileStatic
-class AwsSdkSesEmailDeliveryStatus {
-    public static final int STATUS_DELIVERED = 1
-    public static final int STATUS_BLACKLISTED = -1
-    public static final int STATUS_NOT_DELIVERED = 0
-}
