--- groovy-module/src/main/groovy/grails/plugin/awssdk/ses/UnsupportedAttachmentTypeException.groovy
@@ -1,7 +1,0 @@
-package grails.plugin.awssdk.ses
-
-import groovy.transform.CompileStatic
-
-@CompileStatic
-class UnsupportedAttachmentTypeException extends Exception {
-}
