+++ src/integration-test/groovy/grails/plugin/awssdk/ses/TestInboxConfig.groovy
