+++ src/integration-test/groovy/grails/plugin/awssdk/ses/groovylogo.png
