+++ src/main/groovy/grails/plugin/awssdk/ses/MailRecipient.groovy
